+++ src/BDSColours.cc
@@ -58,6 +58,11 @@ BDSColours::BDSColours()
   // special
   colours["default"]          = new G4Colour(0.9,   0.9,   0.9);   // almost white
   colours["warning"]          = new G4Colour(1,     0.078, 0.576); // hot warning pink
+  
+  // visualisation of trajectories
+  colours["traj_neutral"]     = new G4Colour(0.2, 0.7, 0.0, 0.2); // 0  charge - nicer green 0, semi-transparent as tonnes of photons usually
+  colours["traj_positive"]    = new G4Colour(0.8, 0.0, 0.0);
+  colours["traj_negative"]    = new G4Colour(0.0, 0.2, 0.9);
 
   // tunnel
   colours["tunnel"]           = new G4Colour(0.545, 0.533, 0.470); // a nice grey